+++ lib/src/options/options.dart
@@ -2,6 +2,7 @@
 library;
 
 export 'src/http_base_options.dart';
+export 'src/http_client_options.dart';
 export 'src/http_options.dart';
 export 'src/http_options_mixin.dart';
 export 'src/http_retry_options.dart';