+++ lib/src/options/src/http_abstract_options.dart
@@ -2,15 +2,15 @@ import '../../coverters/converter.dart';
 import 'http_retry_options.dart';
 
 /// 基础选项
-abstract class HttpAbstractOptions<R extends BaseResp<T>, T> {
+abstract class HttpAbstractOptions {
   // ignore: public_member_api_docs
   HttpAbstractOptions({
-    required this.converter,
+    required this.converterOptions,
     this.retryOptions,
   });
 
-  /// 转换
-  final Converter<R, T>? converter;
+  /// 转换选项
+  final ConverterOptions converterOptions;
 
   /// 重试
   final HttpRetryOptions? retryOptions;